+++ src/index.js
@@ -1,12 +1,13 @@
 // @flow
 import Vue from 'vue'
 
-export const mq = Symbol('mq')
+export const MQ = Symbol('mq')
+export const MQMAP = Symbol('mqueries')
 
 export default (Vue: Vue, options?: Object): void => {
   Object.defineProperty(Vue.prototype, '$mq', ({
     get () {
-      return this[mq].obs
+      return this[MQ]
     }
   }: Object))
 
@@ -14,42 +15,42 @@ export default (Vue: Vue, options?: Object): void => {
     beforeCreate () {
       const isIsolated = this.$options.mq && this.$options.mq.config && this.$options.mq.config.isolated
       const isRoot = this === this.$root
-
-      this[mq] = {}
-      this[mq].obs = isIsolated || isRoot || !this.$parent ? {} : this.$parent.$mq
+      const inherited = this.$parent && this.$parent[MQMAP]
+      const inheritedKeys = isIsolated || isRoot || !inherited ? [] : Object.keys(inherited)
 
       if (this.$options.mq) {
-        const observables = Object.keys(this.$options.mq)
-          .filter(k => k !== 'config')
+        this[MQMAP] = {}
+
+        const mergedKeys = new Set(inheritedKeys.concat(
+          Object.keys(this.$options.mq)
+            .filter(k => k !== 'config')
+        ))
+
+        const observed = Array.from(mergedKeys)
           .reduce((obs, k) => {
-            const mql = window.matchMedia(this.$options.mq[k])
-            Object.defineProperty(obs, k, ({
-              enumerable: true,
-              configurable: true,
-              get () {
-                return mql.matches
-              }
-            }: Object))
-            /* Below for testing and debugging */
-            if (process.env.NODE_ENV !== 'production') {
-              this[mq][`_${k}`] = mql.media
-            }
+            const ownQuery = this.$options.mq[k]
+            const mql = ownQuery ? window.matchMedia(ownQuery) : inherited[k]
+            mql.addListener(e => { obs[k] = e.matches })
+
+            obs[k] = mql.matches
+            this[MQMAP][k] = mql
             return obs
           }, {})
 
-        if (!isIsolated) {
-          this[mq].obs = {
-            ...this[mq].obs,
-            ...observables
-          }
-        }
-
-        Object.defineProperty(this[mq].obs, 'all', ({
+        Object.defineProperty(observed, 'all', ({
+          enumerable: true,
+          configurable: true,
           get () {
             return Object.keys(this)
+              .filter(k => k !== 'all')
               .filter(k => this[k])
           }
         }: Object))
+
+        Vue.util.defineReactive(this, MQ, observed)
+      } else if (inherited) {
+        this[MQMAP] = inherited
+        Vue.util.defineReactive(this, MQ, this.$parent[MQ])
       }
     }
   })