+++ src/index.test.js
@@ -1,12 +1,11 @@
-import Vue from 'vue'
-import plugin, {mq} from './index.js'
+import Vue from 'vue/dist/vue.js'
+import plugin, {MQ} from './index.js'
 import 'jasmine-expect'
 import matchMediaMock from 'match-media-mock'
 
-global.window = {}
-global.window.matchMedia = matchMediaMock.create()
-global.window.resizeTo = (x, y) => {
-  global.window.matchMedia.setConfig({
+window.matchMedia = matchMediaMock.create()
+window.resizeTo = (x, y) => {
+  window.matchMedia.setConfig({
     type: 'screen',
     width: x
   })
@@ -16,7 +15,7 @@ beforeAll(() => {
   Vue.use(plugin)
 })
 beforeEach(() => {
-  global.window.resizeTo(1400, 0)
+  window.resizeTo(1400, 0)
 })
 
 const rootOpts = {
@@ -25,10 +24,6 @@ const rootOpts = {
 }
 
 describe('The plugin', () => {
-  it('makes an $mq getter available in the Vue instance', () => {
-    const vm = new Vue()
-    expect(vm.$mq).toBeInstanceOf(Object)
-  })
   it('accepts media query options on the root Vue instance', () => {
     const vm = new Vue({
       mq: rootOpts
@@ -37,19 +32,16 @@ describe('The plugin', () => {
     expect(vm.$mq).toHaveProperty('desktop')
   })
   it('provides an "all" property on the $mq object', () => {
-    global.window.resizeTo(700, 0)
     const vm = new Vue({
       mq: {
-        phone: '(max-width: 768px)',
-        tablet: '(max-width: 1024px)'
-      },
-      render (h) {
-        return h()
+        phone: '(max-width: 728px)',
+        tablet: '(min-width: 728px)',
+        desktop: '(min-width: 1024px)'
       }
     })
-    vm.$mount()
     expect(vm.$mq).toHaveProperty('all')
-    expect(vm.$mq.all.join(' ')).toEqual('phone tablet')
+    expect(vm.$mq.all).toBeArrayOfSize(2)
+    expect(vm.$mq.all).not.toEqual(expect.arrayContaining(['phone']))
   })
   it('implicitly provides media query options to child components', () => {
     const child = Vue.component('child', Vue.extend({
@@ -101,7 +93,7 @@ describe('The plugin', () => {
     expect(vmchild.$mq).toHaveProperty('tablet', false)
     expect(vmchild.$mq).toHaveProperty('desktop', true)
 
-    expect(vmchild[mq]._tablet).toEqual(vmchild[mq]._phone)
+    expect(vmchild[MQ]._tablet).toEqual(vmchild[MQ]._phone)
   })
   it('allows a child component to declare an isolated scope', () => {
     const child = Vue.component('child', Vue.extend({
@@ -130,5 +122,126 @@ describe('The plugin', () => {
     vm.$mount()
     const vmchild = vm.$children[0]
     expect(vmchild.$mq).not.toHaveProperty('desktop')
+  })
+})
+
+describe('In the browser context', () => {
+  beforeEach(() => {
+    document.body.appendChild(document.createElement('main'))
+  })
+  afterEach(() => {
+    document.body.removeChild(document.querySelector('#test'))
+  })
+  test('$mq properties are reactive on the root instance', async () => {
+    const vm = new Vue({
+      mq: rootOpts,
+      template: '<div id="test" :class="{tablet: $mq.tablet}"></div>'
+    })
+    vm.$mount('main')
+    expect(document.getElementById('test').classList.contains('tablet')).toBe(false)
+    window.resizeTo(1000, 0)
+    await vm.$nextTick()
+    expect(document.getElementById('test').classList.contains('tablet')).toBe(true)
+  })
+  test('$mq.all is reactive', async () => {
+    const vm = new Vue({
+      mq: {
+        phone: '(max-width: 728px)',
+        tablet: '(min-width: 700px)',
+        desktop: '(min-width: 1024px)'
+      },
+      template: '<div id="test" :class="$mq.all"></div>'
+    })
+    vm.$mount('main')
+    const classList = document.getElementById('test').classList
+    expect(classList.contains('phone')).toBe(false)
+    expect(classList.contains('tablet')).toBe(true)
+    expect(classList.contains('desktop')).toBe(true)
+    window.resizeTo(710, 0)
+    await vm.$nextTick()
+    expect(classList.contains('phone')).toBe(true)
+    expect(classList.contains('tablet')).toBe(true)
+    expect(classList.contains('desktop')).toBe(false)
+  })
+  test('an inheriting child instance get reactive properties', async () => {
+    const child = Vue.component('child', Vue.extend({
+      template: '<div id="test" :class="{tablet: $mq.tablet}"></div>'
+    }))
+    const vm = new Vue({
+      mq: rootOpts,
+      components: {
+        child
+      },
+      render (h) {
+        return h(child)
+      }
+    })
+    vm.$mount('main')
+    expect(document.getElementById('test').classList.contains('tablet')).toBe(false)
+    window.resizeTo(1000, 0)
+    await vm.$nextTick()
+    expect(document.getElementById('test').classList.contains('tablet')).toBe(true)
+  })
+  test('and when an inheritor overrides root properties', async () => {
+    const child = Vue.component('child', Vue.extend({
+      template: '<div id="test" :class="$mq.all"></div>',
+      mq: {
+        phone: '(max-width: 760px)',
+        tablet: '(max-width: 760px)'
+      }
+    }))
+    const vm = new Vue({
+      mq: rootOpts,
+      components: {
+        child
+      },
+      data: {
+        name: 'root'
+      },
+      render (h) {
+        return h(child)
+      }
+    })
+    vm.$mount('main')
+    const classList = document.getElementById('test').classList
+    expect(classList.contains('phone')).toBe(false)
+    expect(classList.contains('tablet')).toBe(false)
+    expect(classList.contains('desktop')).toBe(true)
+    window.resizeTo(710, 0)
+    await vm.$nextTick()
+    expect(classList.contains('phone')).toBe(true)
+    expect(classList.contains('tablet')).toBe(true)
+    expect(classList.contains('desktop')).toBe(false)
+  })
+  test('or isolates itself', async () => {
+    const child = Vue.component('child', Vue.extend({
+      template: '<div id="test" :class="$mq.all"></div>',
+      mq: {
+        phone: '(max-width: 760px)',
+        tablet: '(max-width: 760px)',
+        config: {
+          isolated: true
+        }
+      }
+    }))
+    const vm = new Vue({
+      mq: rootOpts,
+      components: {
+        child
+      },
+      data: {
+        name: 'root'
+      },
+      render (h) {
+        return h(child)
+      }
+    })
+    vm.$mount('main')
+    const classList = document.getElementById('test').classList
+    expect(classList.length).toBe(0) // "desktop" would have been hit if this weren't isolated
+    window.resizeTo(710, 0)
+    await vm.$nextTick()
+    expect(classList.contains('phone')).toBe(true)
+    expect(classList.contains('tablet')).toBe(true)
   })
 })