+++ CHANGELOG.md
@@ -6,10 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## 0.3.0 - 2017-06-27
+### Added
+- Reactive properties enabled on $mq object
+
+### Fixed
+- Initial form of option merge was naive and didn't support reactivity
+
 ## 0.2.0 - 2017-06-26
 ### Added
-- Proper matchMedia mock for unit testing
 - Convenience property "all" on $mq object returns array of all matched query keys
+
+### Fixed
+- Added a proper matchMedia mock for testing
 
 ## 0.1.0 - 2017-06-25
 ### Added